chore: create separate class for ScreenEvent class and add documentation
